Unit, traits and control
Add a unit, a basic set of properties (id and position).
Add a trait, a pair of properties and commands that extends unit or other traits.
Add a control block, a set of operations upon a unit with specific properties. Control block is union of all traits' commands.
	`auto c = control_block_t<...>{unit}; auto f = c.move(...).work(...).rest(...).finish();`

Add a control module that processes a queue of control blocks. Control module is a generalisation over Control block that is generalization over Operation.
```
	auto c = control_module_t<...>{unit};
	auto f = c.start().move(...)....finish(); // first block
	auto f2 = c.start().work(...).eat(...).sleep(...).finish(); // second block, executed after first.
	auto f3 = c.idle(); // Idle state, if all blocks have completed
```